refactor(mcp-server): centralize TIMELESS_DATE date-validation primitives
The charges and reports tools each defined an identical TIMELESS_DATE zod
schema, plus duplicate DAY_MS and parseCalendarDate helpers. Extract these
shared date-validation primitives into a single tools/dates.ts module and
import them in both tools. Behavior is unchanged; the per-tool
assertDateRange functions stay put since their range rules differ.
